app/requirements.txt
loguru
fastapi
uvicorn
joblib
scikit-learn
opentelemetry-api==1.19.0
opentelemetry-sdk==1.19.0
opentelemetry-instrumentation-asgi==0.40b0
opentelemetry-instrumentation-fastapi==0.40b0
opentelemetry-instrumentation-requests==0.40b0
opentelemetry-instrumentation-logging==0.40b0
opentelemetry-exporter-otlp-proto-grpc==1.19.0
opentelemetry-exporter-prometheus==1.12.0rc1
prometheus-client==0.17.1
